Check if dir exists

+++ 0 Setup/Setup.cs
@@ -166,8 +166,11 @@ public partial class Setup : Node
 
     void DeleteDirectoryIfEmpty(string path)
     {
-        GU.DeleteEmptyFolders(path);
-        GU.DeleteEmptyFolder(path);
+        if (Directory.Exists(path))
+        {
+            GU.DeleteEmptyFolders(path);
+            GU.DeleteEmptyFolder(path);
+        }
     }
 
     /// <summary>
